[Refactoring] Move SimplifyNumberLiteral to its own file
